+++ includes/DataLayer/Bet.h
@@ -1,10 +1,12 @@
 #pragma once
 
+#include <string>
+
 namespace DataLayer
 {
 	struct Bet
 	{
 		double amount = 0;
-		int handIndex = 0;
+		size_t handIndex = 0;
 	};
 }
